Use variables, "all" and "clean" targets, and compiler flags
The CC variable contains the compiler to use, CFLAGS contains compiler flags (I used -std=c11 which specifies the C standard, and -O which optimises the code), and OBJS contains the object files to link into the executable. I changed the name of the "main" target to "pingyclash" because it's a better name than "main". I also changed some behavior in the targets, $@ means the name of the target, and $< means the first dependency of the target. I also added an "all" target which is considered good practice, and a "clean" target you can use to remove all compiled code (you can use it if you wont to recompile).